Update defaultWhitelistArsenal.json
Removed poor quality Unsung M2
Removed unnecessary L1A1 variants
Added equipment in the arsenal box to the actual arsenal to make loading of kits easier. If unicorn equipment is overused we can restrict later. Only thing left in the arsenal box is 30 round magazines. Did not add Stoner to the arsenal as it was never used.